+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [#60](https://github.com/LiamMartens/sanity-plugin-intl-input/issues/60) Improved publish with I18n action and expose "updateIntlFieldsForDocument" action (ie to be used with custom publish action)
 * [#72](https://github.com/LiamMartens/sanity-plugin-intl-input/issues/72) Updated custom delete action
 * It is now possible to define a custom loader and reload function for loading languages.
+* It is now possible to import all custom actions from the plugin for re-use or customization.
 
 ## v5.2.1
 * [#58](https://github.com/LiamMartens/sanity-plugin-intl-input/issues/58) Only disable publishing when there are errors (no warnings)